+++ accuracode/__init__.py
@@ -1,6 +1,6 @@
 import os
 
-__VERSION__ = "1.1.0"
+__VERSION__ = "1.2.0"
 __version__ = __VERSION__
 
 ASSAY_DICT = {