+++ accuracode/tools/__init__.py
@@ -2,7 +2,6 @@
 __PATTERN_DICT__ = {'accuracode96': 'C9U12T18',
                     'accuracode384': 'C9U12T18',
                     'customized': None,
-                    'auto': None,
                     }