+++ docs/CHANGELOG.md
@@ -1,3 +1,12 @@
+## [v1.2.0] - 2023-07-27
+ ### `accura`
+ - Linker file not required if `L` is absent in Pattern
+ - Customized white list can have two columns
+ ### General improvements
+ - Update installation of accuracode
+ - Update software version (STAR/picard/cutadapt)
+
+
 ## [v1.1.0] - 2023-03-17
  ### `accura` 
  - Change barcode sequence to oligo ID in report and count matrix (Not performed in customized chemistry)